backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-slugify>=8.0.1

# Security
python-multipart>=0.0.6

# AI/LLM (Phase 1)
anthropic>=0.40.0

# Content Extraction
httpx>=0.27.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1

# SSE Streaming
sse-starlette>=2.0.0

# RAG Preparation (Phase 2)
# pgvector>=0.2.0
# fastembed>=0.3.0
